Adding deno implementation
