--- .github/workflows/aunty-bot.yml
@@ -1,32 +1,0 @@
-#- cron: "30 23,5 * * 0-4"
-name: Show Who's running Story Lab standup
-
-on:
-  workflow_dispatch:
-  #schedule:
-  #  - cron: "30 23,5 * * 0-4"
-
-jobs:
-  build-and-choose:
-
-    runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        node-version: [18.x]
-        # See supported Node.js release schedule at https://nodejs.org/en/about/releases/
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Use Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@v3
-      with:
-        node-version: ${{ matrix.node-version }}
-        cache: 'npm'
-    - run: npm ci
-    - run: npm run build --if-present
-    - name: create env file
-      run: |
-        touch .env
-        echo SLACK_URL=${{ secrets.SLACK_URL }} >> .env
-    - run: npm run start
