+++ .prettierrc.toml
@@ -1,0 +1,1 @@
+arrowParens = "avoid"
