--- .prettierrc.yaml
@@ -1,10 +1,0 @@
----
-trailingComma: none
-bracketSpacing: true
-useTabs: false
-tabWidth: 2
-semi: true
-singleQuote: false
-arrowParens: avoid
-bracketSameLine: true
-printWidth: 100
