--- README.md
@@ -1,2 +1,0 @@
-# story-lab-standup-bot
-Determine who will run Story Lab standups
