+++ deno.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "noImplicitAny": false
+  },
+  "deploy": {
+    "project": "f9405a8a-75c7-42e8-badd-5308e4129262",
+    "exclude": [],
+    "include": [],
+    "entrypoint": "main.ts"
+  }
+}
