--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "storylab-slack-bot",
-  "version": "1.0.0",
-  "description": "",
-  "main": "dist/index.js",
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/index.js",
-    "dev": "ts-node src/index.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "async-retry": "^1.3.3",
-    "await-to-js": "^3.0.0",
-    "axios": "^1.1.3",
-    "dotenv": "^16.0.3",
-    "ts-node": "^10.9.1"
-  },
-  "devDependencies": {
-    "@types/async-retry": "^1.4.5",
-    "typescript": "^4.8.4"
-  }
-}
