--- tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitAny": false,
-    "module": "commonjs",
-    "esModuleInterop": true,
-    "target": "es6",
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "outDir": "dist"
-  },
-  "lib": ["es2015"]
-}
